Set minimum time-lapse duration to be 5 min

+++ web/app/tasks.py
@@ -16,38 +16,42 @@ from lib.file_storage import list_file_obj, retrieve_to_file_obj, save_file_obj
 
 @shared_task(acks_late=True, bind=True, autoretry_for=(Exception,), retry_kwargs={'max_retries': 3}, retry_backoff=True)
 def compile_timelapse(self, print_id):
-    pprint = Print.objects.get(id=print_id)
-    end_time = pprint.finished_at or pprint.cancelled_at
+    print = Print.objects.get(id=print_id)
+    end_time = print.finished_at or print.cancelled_at
 
-    to_dir = os.path.join(tempfile.gettempdir(), str(pprint.id))
+    if (end_time - print.started_at).total_seconds() < settings.TIMELAPSE_MINIMUM_SECONDS:
+        print.delete()
+        return
+
+    to_dir = os.path.join(tempfile.gettempdir(), str(print.id))
     shutil.rmtree(to_dir, ignore_errors=True)
     os.mkdir(to_dir)
 
-    print_pics = filter_pics_by_start_end(list_file_obj('raw/{}/'.format(pprint.printer.id), settings.PICS_CONTAINER), pprint.started_at, end_time)
+    print_pics = filter_pics_by_start_end(list_file_obj('raw/{}/'.format(print.printer.id), settings.PICS_CONTAINER), print.started_at, end_time)
     print_pics.sort()
     if print_pics:
         local_pics = download_files(print_pics, to_dir)
         last_pic = local_pics[-1]
-        mp4_filename = '{}.mp4'.format(pprint.id)
+        mp4_filename = '{}.mp4'.format(print.id)
         output_mp4 = os.path.join(to_dir, mp4_filename)
         subprocess.run('ffmpeg -y -pattern_type glob -i {}/*.jpg -c:v libx264 -vf fps=30 -pix_fmt yuv420p {}'.format(last_pic.parent, output_mp4).split(' '), check=True)
-        shutil.copyfile(last_pic, os.path.join(to_dir, '{}.jpg'.format(pprint.id)))
+        shutil.copyfile(last_pic, os.path.join(to_dir, '{}.jpg'.format(print.id)))
 
         with open(output_mp4, 'rb') as mp4_file:
             _, mp4_file_url = save_file_obj('private/{}'.format(mp4_filename), mp4_file, settings.TIMELAPSE_CONTAINER)
         with open(last_pic, 'rb') as poster_file:
-            _, poster_file_url = save_file_obj('private/{}_poster.jpg'.format(pprint.id), poster_file, settings.TIMELAPSE_CONTAINER)
+            _, poster_file_url = save_file_obj('private/{}_poster.jpg'.format(print.id), poster_file, settings.TIMELAPSE_CONTAINER)
 
-        pprint.video_url = mp4_file_url
-        pprint.poster_url = poster_file_url
-        pprint.save()
+        print.video_url = mp4_file_url
+        print.poster_url = poster_file_url
+        print.save()
 
     # build tagged timelapse
-    print_pics = filter_pics_by_start_end(list_file_obj('tagged/{}/'.format(pprint.printer.id), settings.PICS_CONTAINER), pprint.started_at, end_time)
+    print_pics = filter_pics_by_start_end(list_file_obj('tagged/{}/'.format(print.printer.id), settings.PICS_CONTAINER), print.started_at, end_time)
     print_pics.sort()
     if print_pics:
         local_pics = download_files(print_pics, to_dir)
-        mp4_filename = '{}_tagged.mp4'.format(pprint.id)
+        mp4_filename = '{}_tagged.mp4'.format(print.id)
         output_mp4 = os.path.join(to_dir, mp4_filename)
         subprocess.run('ffmpeg -y -pattern_type glob -i {}/*.jpg -c:v libx264 -vf fps=30 -pix_fmt yuv420p {}'.format(local_pics[0].parent, output_mp4).split(' '), check=True)
         with open(output_mp4, 'rb') as mp4_file:
@@ -66,11 +70,11 @@ def compile_timelapse(self, print_id):
         preidction_json_io = io.BytesIO()
         preidction_json_io.write(json.dumps(preidction_json).encode('UTF-8'))
         preidction_json_io.seek(0)
-        _, json_url = save_file_obj('private/{}_p.json'.format(pprint.id), preidction_json_io, settings.TIMELAPSE_CONTAINER)
+        _, json_url = save_file_obj('private/{}_p.json'.format(print.id), preidction_json_io, settings.TIMELAPSE_CONTAINER)
 
-        pprint.tagged_video_url = mp4_file_url
-        pprint.prediction_json_url = json_url
-        pprint.save()
+        print.tagged_video_url = mp4_file_url
+        print.prediction_json_url = json_url
+        print.save()
 
     shutil.rmtree(to_dir, ignore_errors=True)
